+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.9.1
+* fix `if x=True` bug in extract
+* fix `ValueError` in `sql_str_to_dict` when using non-space whitespace
+
+# 0.9.0
+* + boilerplate for `extract --link`
+* more `lru_cache` usage
+* `select` allows more SQL commands
+* now (theoretically) ignoring duplicate files when same multihashes (#41)
+
 # 0.8.8
 * add `lru_cache`ing to reduce loads on metadata db